fixes "no active program" issue
Colorwheel, like Flywheel, changed the OpenGL program manually.
This is fine for Flywheel as the correct state is restored
using the GlStateTracker.

However Iris, being smart, adds a check to prevent rebinding
the program if it's already in use. Because Colorwheel didn't
contributed to GlStateManager, the in-memory state and the real state
mismatched and this prevented the program restore.